testing/messing with version overides [skip ci]

+++ src/BladeExtensions.php
@@ -5,11 +5,12 @@
  * The license can be found in the package and online at https://radic.mit-license.org.
  *
  * @copyright Copyright 2017 (c) Robin Radic
- * @license   https://radic.mit-license.org The MIT License
+ * @license https://radic.mit-license.org The MIT License
  */
 
 namespace Radic\BladeExtensions;
 
+use Closure;
 use Illuminate\Filesystem\Filesystem;
 use Illuminate\View\Compilers\BladeCompiler;
 
@@ -27,15 +28,6 @@ class BladeExtensions
     /** @var \Illuminate\Filesystem\Filesystem */
     protected $fs;
 
-    /** @var bool */
-    protected $hooked = false;
-
-    /** @var string */
-    protected $mode;
-
-    /** @var \Closure|null */
-    protected $customModeHandler;
-
     /** @var string */
     protected $cachePath;
 
@@ -50,9 +42,9 @@ class BladeExtensions
     public function __construct(DirectiveRegistry $directives, HelperRepository $helpers)
     {
         $this->directives = $directives;
-        $this->helpers = $helpers;
-        $this->fs = new Filesystem;
-        $this->cachePath = storage_path('blade-extensions');
+        $this->helpers    = $helpers;
+        $this->fs         = new Filesystem;
+        $this->cachePath  = storage_path('blade-extensions');
     }
 
     /**
@@ -61,8 +53,8 @@ class BladeExtensions
      */
     protected function getCompiler()
     {
-        if (! isset($this->compiler)) {
-            if ($this->fs->exists($this->cachePath) === false) {
+        if ( !isset($this->compiler) ) {
+            if ( $this->fs->exists($this->cachePath) === false ) {
                 $this->fs->makeDirectory($this->cachePath);
             }
             $this->compiler = new BladeCompiler($this->fs, $this->cachePath);
@@ -71,11 +63,22 @@ class BladeExtensions
         return $this->compiler;
     }
 
-    public function compileString($string, array $vars = [], $tmpfile = false)
+    /**
+     * Compile blade syntax to string
+     *
+     * @param string $string String with blade syntax to compile
+     * @param array  $vars   Optional variables
+     *
+     * @return string
+     */
+    public function compileString($string, array $vars = [])
     {
-        $fileName = uniqid('compileString', true).'.php';
-        $filePath = $this->cachePath.DIRECTORY_SEPARATOR.$fileName;
-        $string = $this->getCompiler()->compileString($string);
+        if ( empty($vars) ) {
+            return $this->getCompiler()->compileString($string);
+        }
+        $fileName = uniqid('compileString', true) . '.php';
+        $filePath = $this->cachePath . DIRECTORY_SEPARATOR . $fileName;
+        $string   = $this->getCompiler()->compileString($string);
         $this->fs->put($filePath, $string);
         $compiledString = $this->getCompiledContent($filePath, $vars);
         $this->fs->delete($filePath);
@@ -85,7 +88,7 @@ class BladeExtensions
 
     protected function getCompiledContent($filePath, array $vars = [])
     {
-        if (is_array($vars) && ! empty($vars)) {
+        if ( is_array($vars) && !empty($vars) ) {
             extract($vars, EXTR_OVERWRITE);
         }
         ob_start();
@@ -95,6 +98,31 @@ class BladeExtensions
 
         return $var;
     }
+
+    /**
+     * pushToStack method
+     *
+     * @param string          $stackName   The name of the stack
+     * @param string|string[] $targetViews The view which contains the stack
+     * @param string|Closure  $content     the content to push
+     *
+     * @return $this
+     */
+    public function pushToStack($stackName, $targetViews, $content)
+    {
+        $targetViews = is_array($targetViews) ? $targetViews : [ $targetViews ];
+        foreach ( $targetViews as $targetView ) {
+            app()->make('events')->listen('composing: ' . $targetView, function ($view) use ($stackName, $content) {
+                $content = $content instanceof Closure ? $content($view) : $content;
+                if ( method_exists($view, 'getFactory') && method_exists($view->getFactory(), 'startPush') ) {
+                    $view->getFactory()->startPush($stackName, $content);
+                }
+            });
+        }
+
+        return $this;
+    }
+
 
     /**
      * @return bool
@@ -118,18 +146,5 @@ class BladeExtensions
     public function getHelpers()
     {
         return $this->helpers;
-    }
-
-    /**
-     * @param string $mode
-     */
-    public function setMode($mode)
-    {
-        $this->mode = $mode;
-    }
-
-    public function setCustomModeHandler($handler)
-    {
-        $this->customModeHandler = $handler;
     }
 }
